move XDSRequestHash/SetXDSRequestHash to internal/xds
