+++ balancer/ringhash/picker.go
@@ -25,6 +25,7 @@ import (
 	xxhash "github.com/cespare/xxhash/v2"
 	"google.golang.org/grpc/balancer"
 	"google.golang.org/grpc/connectivity"
+	"google.golang.org/grpc/internal/xds"
 	"google.golang.org/grpc/metadata"
 )
 
@@ -55,7 +56,7 @@ func (p *picker) Pick(info balancer.PickInfo) (balancer.PickResult, error) {
 	var requestHash uint64
 	if p.requestHashHeader == "" {
 		var ok bool
-		if requestHash, ok = XDSRequestHash(info.Ctx); !ok {
+		if requestHash, ok = xds.GetXDSRequestHash(info.Ctx); !ok {
 			return balancer.PickResult{}, fmt.Errorf("ringhash: expected xDS config selector to set the request hash")
 		}
 	} else {
